cleaning up appswitcher and sidebar
