+++ lib/components/formik/Select.js
@@ -49,11 +49,6 @@ const SelectField = forwardRef((props, ref) => {
   );
 });
 
-SelectField.defaultProps = {
-  options: [],
-  name: "",
-};
-
 SelectField.propTypes = {
   /**
    * The name of the select input.
