+++ lib/components/layouts/AppSwitcher/index.js
@@ -2,28 +2,19 @@ import React, { useRef } from "react";
 import { AnimatePresence, motion } from "framer-motion";
 import { useHotkeys } from "react-hotkeys-hook";
 import classnames from "classnames";
-
+import PropTypes from "prop-types";
 import Body from "./Body";
 import { Portal, Backdrop } from "../../../atoms";
 import { useOnClickOutside } from "../../../utils";
 import {
   TRANSITION,
   ANIMATION_VARIANTS,
-  APPSWITCHER_PROPTYPES,
+  ENVS,
   SIZES,
+  getSubdomain,
 } from "./constant";
 
 const noop = () => {};
-
-const getSubdomain = () => {
-  let host = window.location.host;
-  let parts = host.split(".");
-  let subdomain = "";
-  if (parts.length >= 3) {
-    subdomain = parts[0];
-  }
-  return subdomain;
-};
 
 const AppSwitcher = ({
   size = "sm",
@@ -100,6 +91,54 @@ const AppSwitcher = ({
   );
 };
 
-AppSwitcher.propTypes = APPSWITCHER_PROPTYPES;
-
+AppSwitcher.propTypes = {
+  /**
+   * Size of the AppSwitcher
+   */
+  size: PropTypes.oneOf(Object.values(SIZES)),
+  /**
+   * To specify whether the AppSwitcher is open or not
+   */
+  isOpen: PropTypes.bool,
+  /**
+   * To provide a callback function on close of the AppSwitcher
+   */
+  onClose: PropTypes.func,
+  /**
+   * To specify whether the AppSwitcher is loading or not
+   */
+  loading: PropTypes.bool,
+  /**
+   * To provide additional classNames to the AppSwitcher
+   */
+  className: PropTypes.string,
+  /**
+   * To close AppSwitcher on esc key press
+   */
+  closeOnEsc: PropTypes.bool,
+  /**
+   * To close AppSwitcher on outside click
+   */
+  closeOnOutsideClick: PropTypes.bool,
+  /**
+   * Name of the active application
+   */
+  activeApp: PropTypes.string,
+  /**
+   * Organization subdomain name
+   */
+  subdomain: PropTypes.string,
+  /**
+   * List of names of enabled neetoApps
+   */
+  neetoApps: PropTypes.arrayOf(PropTypes.string),
+  /**
+   * List of names of recently created neetoApps
+   */
+  recentApps: PropTypes.arrayOf(PropTypes.string),
+  /**
+   * To specify the environment of the App
+   */
+  environment: PropTypes.oneOf(Object.values(ENVS)).isRequired,
+};
 export default AppSwitcher;
